pyproject.toml
[project]
name = "fin-toolkit"
version = "0.1.0"
description = "Protocol-first financial analysis toolkit with MCP server for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.0",
    "ta>=0.11",
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
fin-toolkit = "fin_toolkit.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.10",
    "ruff>=0.4",
    "types-pyyaml>=6.0",
    "types-beautifulsoup4>=4.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "live: marks tests that hit real external APIs (deselect with '-m \"not live\"')",
]

[tool.coverage.run]
source = ["fin_toolkit"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
